erts: Disable unsafe optimization in bs_append
Imagine that we execute the following code, where `X` and `Y` are
integers:

```
A = <<0, X>>,
B = <<A/bits, Y:X>>.
```

The first line sets up `erts_current_bin`, `erts_bin_offset`, etc
because it obviously should.

However, the second line does not do so if X is zero, as
`build_size_in_bits` being zero short-circuits the relevant logic.

This leaves a dangling pointer that, when we execute
`erts_new_bs_put_integer` on the second line, we can land in a code
path that ostensibly does not modify the buffer, but may
nevertheless do a dummy read/write that is masked off, which can
race with a write to the same address on another scheduler. For
example, this can occur if there is a garbage collection between
`A` and `B`, where the old heap that `A` lived on is picked up by
another process before `B` executes.

Note that there does not need to be an actual link between `A` and
`B` here: all that is necessary is that an expression like
`<<A/bits, Y:X>>` is executed where `X` is zero. This issue was
mostly benign before 9256aad02f77bc4ca8c2ed5a4b41d5cad834fad1 that
introduced an read-modify-write in certain cases that would be okay
if `erts_current_bin` and `erts_bin_offset` had been set correctly.

While we could fix the read-modify-write issue separately, the
invariant that `erts_current_bin` and `erts_bin_offset` are up to
date has been broken, and it's difficult to tell whether other
parts may suffer the same problem. Therefore, we will disable this
optimization and reintroduce it again safely at a later time.